Rename namespace lexer to lex and change CMake and directory structure
