+++ src/CMakeLists.txt
@@ -1,6 +1,6 @@
 # Create object files for lexer
 add_library(LEXER_LIB_OBJECTS OBJECT
-    ${CMAKE_CURRENT_SOURCE_DIR}/core/lexer/lexer.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/core/lex/lexer.cpp
     )
 target_include_directories(LEXER_LIB_OBJECTS PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}
